+++ Element/AbstractElement.php
@@ -50,7 +50,24 @@ abstract class AbstractElement implements GenerableInterface
      */
     public static function nameIsValid($name)
     {
-        return preg_match('/^[a-zA-Z_\x7f-\xff][a-zA-Z0-9_\x7f-\xff]*$/D', $name) === 1;
+        return preg_match('/^[a-zA-Z_][a-zA-Z0-9_]*$/D', $name) === 1;
+    }
+    /**
+     * @param mixed $string
+     * @param bool $checkName
+     * @return bool
+     */
+    public static function stringIsValid($string, $checkName = true)
+    {
+        return (is_string($string) && !empty($string) && (!$checkName || self::nameIsValid($string)));
+    }
+    /**
+     * @param mixed $object
+     * @return bool
+     */
+    public static function objectIsValid($object, $checkClass = null)
+    {
+        return (is_object($object) && ($checkClass === null || get_class($object) === $checkClass));
     }
     /**
      * @see \WsdlToPhp\PhpGenerator\Element\GenerableInterface::toString()
@@ -59,40 +76,79 @@ abstract class AbstractElement implements GenerableInterface
      */
     public function toString($indentation = null)
     {
-        $lines = array();
-        $declaration = $this->getPhpDeclaration();
-        if (!empty($declaration)) {
-            $lines = array(
-                $this->getIndentedString($declaration, $indentation),
-            );
-        }
-        $before = $this->getContextualLineBeforeChildren($indentation);
-        if (!empty($before)) {
-            $lines[] = $before;
-        }
+        $lines = array(
+            $this->getToStringDeclaration($indentation),
+            $this->getToStringBeforeChildren($indentation),
+        );
         foreach ($this->getChildren() as $child) {
             $lines[] = $this->getChildContent($child, $indentation + ($this->useBracketsForChildren() ? 1 : 0));
         }
-        $after = $this->getContextualLineAfterChildren($indentation);
-        if (!empty($after)) {
-            $lines[] = $after;
-        }
-        return implode(self::BREAK_LINE_CHAR, $lines);
+        $lines[] = $this->getToStringAfterChildren($indentation);
+        return implode(self::BREAK_LINE_CHAR, self::cleanArrayToString($lines));
     }
     /**
+     * @param int $indentation
+     * @return string|null
+     */
+    private function getToStringDeclaration($indentation = null)
+    {
+        $declaration = $this->getPhpDeclaration();
+        if (!empty($declaration)) {
+            return $this->getIndentedString($declaration, $indentation);
+        }
+        return null;
+    }
+    /**
+     * @param string $indentation
+     * @return string|null
+     */
+    private function getToStringBeforeChildren($indentation = null)
+    {
+        $before = $this->getContextualLineBeforeChildren($indentation);
+        if (!empty($before)) {
+            return $before;
+        }
+        return null;
+    }
+    /**
+     * @param string $indentation
+     * @return string|null
+     */
+    private function getToStringAfterChildren($indentation = null)
+    {
+        $after = $this->getContextualLineAfterChildren($indentation);
+        if (!empty($after)) {
+            return $after;
+        }
+        return null;
+    }
+    /**
+     * @param array $array
+     * @return array
+     */
+    private static function cleanArrayToString($array)
+    {
+        $newArray = array();
+        foreach ($array as $line) {
+            if ($line !== null) {
+                $newArray[] = $line;
+            }
+        }
+        return $newArray;
+    }
+    /**
+     * @throws \InvalidArgumentException
      * @param string|AbstractElement $child
      * @param int $indentation
-     * @throws \InvalidArgumentException
      * @return string
      */
     protected function getChildContent($child, $indentation = null)
     {
+        $content = '';
         if (is_string($child)) {
             $content = $this->getIndentedString($child, $indentation);
         } elseif ($child instanceof AbstractElement) {
             $content = $child->toString($indentation === null ? $this->getIndentation() : $indentation);
-        } else {
-            throw new \InvalidArgumentException(sprintf('Child\'s content could not be generated for: %s:%s', gettype($child), is_object($child) ? get_class($child) : 'not an object'));
         }
         return $content;
     }
@@ -104,13 +160,13 @@ abstract class AbstractElement implements GenerableInterface
         return sprintf('%s', $this->getName());
     }
     /**
-     * @param mixed $child
      * @throws \InvalidArgumentException
+     * @param mixed $child
      * @return AbstractElement
      */
     public function addChild($child)
     {
-        if (!$this->childrenIsValid($child)) {
+        if (!$this->childIsValid($child)) {
             $types = $this->getChildrenTypes();
             if (empty($types)) {
                 throw new \InvalidArgumentException('This element does not accept any child element');
@@ -125,13 +181,13 @@ abstract class AbstractElement implements GenerableInterface
      * @param mixed $child
      * @return bool
      */
-    protected function childrenIsValid($child)
+    protected function childIsValid($child)
     {
         $valid = false;
         $authorizedTypes = $this->getChildrenTypes();
         if (!empty($authorizedTypes)) {
             foreach ($authorizedTypes as $authorizedType) {
-                $valid |= (gettype($child) === $authorizedType) || (is_object($child) && get_class($child) === $authorizedType);
+                $valid |= (gettype($child) === $authorizedType) || self::objectIsValid($child, $authorizedType);
             }
         }
         return (bool)$valid;
@@ -256,6 +312,7 @@ abstract class AbstractElement implements GenerableInterface
     }
     /**
      * @param string $string
+     * @param int $indentation
      * @return string
      */
     public function getIndentedString($string, $indentation = null)