+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 CHANGELOG
 =========
 
+0.0.2
+-----
+Code coverage improved
+Several methods have been refactored to minimize them and consolidate them
+Component/PhpInterface has been added
+Component/PhpFile has been cleaned and enhanced
+Component/PhpClass has been cleaned
+Component/AbstractComponent has been enhanced
+
 0.0.1
 -----
 Initial version